Change for path/file .dephpugger.yml preferences
So, first of all, thanks so much for this package. I consider a great idea, I coming from ruby on rails and one tool like byebug was all what I wanted for my web applications in php, most specifically, in Laravel.

Well, this change is just an detail that I have figured when I tried use the preferences in .dephpugger.yml file, related to the path passed at the '-t' parameter. When I used the path param in the .dephpugger.yml, it was duplicated with the variable $projetcPath.

I hope this can be really useful :)

PS: Sorry for my bad english :-P