Remove manually loading of command class files
Remove manually loading of class files and pass this responsability to
composer autoload.

Following SRP and IoC, this commit also remove responsability of
commands to be included on application and pass the responsaility of
instantiate and add to console to `\Dephpug\Dephpugger`.

Now, all console classes and function could be used in development
environment, including automated tests and the bin/debugger is more
cleaner, having the only responsibility of instantiate and init
`\Dephpug\Dephpugger`.